refactor: 全模块结构体字段重排优化内存布局并新增进程优先级接口
主体为内存布局优化重构，附带 utils 模块一个跨平台新接口。

- 全模块（kernel / http / http2 / ws / rpc / mcp / redis / mysql / mongo /
  etcd / ssl / tracing / utils 及全部 C ABI 模块）的 struct / class 成员按访问
  热点与尺寸重排，收敛分散的小标量、指针与 bool 以减少 padding、提升缓存命中率；
  同步更新聚合初始化与构造初始化列表顺序以匹配新声明顺序，消除 -Wreorder。
- c_coro_tcp_bridge.cc 的 CoroTcpOperationBase 用单个 uint64_t m_flags 位标记
  替代 m_finished / m_complete_accepted 两个独立 bool，并提供 finished() /
  completeAccepted() / setFinished() / setCompleteAccepted() 访问器。
- kernel IO 上下文部分 bool / 枚举状态字段统一改为 uint64_t 承载：
  ReadvIOContext / WritevIOContext 的 m_immediate_result、FileWatchIOContext
  的 m_events、SequenceAwaitableBase 的 m_registered。
- utils 新增跨平台进程优先级接口：Process::priority() / setPriority()，
  POSIX 基于 getpriority / setpriority（nice 值 [-20,19]），Windows 映射到
  priority class；新增 ProcessPriorityError 错误枚举与配套错误描述函数，
  错误经 std::expected<T, ProcessPriorityError> 显式传播，errno / GetLastError
  立即转换为具体错误码；module_prelude.hpp 补齐 <cerrno> / <sys/resource.h>，
  新增 test/cpp/utils/t11_platform_process_system.cc 白盒测试。

+++ src/cpp/galay-kernel/concurrency/mpsc_channel.h
@@ -144,7 +144,7 @@ public:
      * @return true 入队成功；false 底层无锁队列拒绝入队
      */
     bool send(T&& value) {
-        uint32_t prevSize = m_size.fetch_add(1, std::memory_order_acq_rel);
+        size_t prevSize = m_size.fetch_add(1, std::memory_order_acq_rel);
         if (!m_queue.enqueue(std::forward<T>(value))) {
             m_size.fetch_sub(1, std::memory_order_acq_rel);
             return false;
@@ -173,7 +173,7 @@ public:
     bool sendBatch(const std::vector<T>& values) requires std::copy_constructible<T> {
         if (values.empty()) return true;
         uint32_t count = static_cast<uint32_t>(values.size());
-        uint32_t prevSize = m_size.fetch_add(count, std::memory_order_acq_rel);
+        size_t prevSize = m_size.fetch_add(count, std::memory_order_acq_rel);
         if (!m_queue.enqueue_bulk(values.data(), values.size())) {
             m_size.fetch_sub(count, std::memory_order_acq_rel);
             return false;
@@ -192,7 +192,7 @@ public:
     bool sendBatch(std::vector<T>&& values) {
         if (values.empty()) return true;
         uint32_t count = static_cast<uint32_t>(values.size());
-        uint32_t prevSize = m_size.fetch_add(count, std::memory_order_acq_rel);
+        size_t prevSize = m_size.fetch_add(count, std::memory_order_acq_rel);
         if (!m_queue.enqueue_bulk(std::make_move_iterator(values.begin()), count)) {
             m_size.fetch_sub(count, std::memory_order_acq_rel);
             return false;
@@ -282,7 +282,7 @@ public:
             values.resize(maxCount);
             size_t count = m_queue.try_dequeue_bulk(values.data() + base, maxCount - base);
             if (count > 0) {
-                m_size.fetch_sub(static_cast<uint32_t>(count), std::memory_order_acq_rel);
+                m_size.fetch_sub(count, std::memory_order_acq_rel);
                 m_waiter_registration.clearPendingWake();
                 values.resize(base + count);
             } else {
@@ -383,7 +383,7 @@ private:
     }
 
 
-    alignas(64) std::atomic<uint32_t> m_size{0};
+    alignas(64) std::atomic<size_t> m_size{0};
     WaitRegistration m_waiter_registration;
     moodycamel::ConcurrentQueue<T> m_queue;
     size_t m_defaultBatchSize{1024};
